+++ libs/charts/src/lib/billboard.spec.ts
@@ -337,6 +337,56 @@ describe('billboard', () => {
       const formattedNumber = (parsed.tooltip.format.value.bind({}))(100, 0.1, undefined, undefined)
       expect(formattedNumber).toEqual('10%')
     })
+    it('add tick config if ticksCount is specified in style', () => {
+      const chartElement = '#foo'
+      const settings: ChartSettings = {
+        data: [{ name: 'Foo', values: [1], axis: 'y2' }],
+        style: {
+          axis: {
+            y2: {
+              show: false,
+              ticksCount: 5
+            },
+          },
+        },
+      }
+      const parsed = createOptions({ settings, chartElement })
+      const expected: Axis = {
+        y2: {
+          show: false,
+          tick: {
+            count: 5,
+            stepSize: null
+          }
+        },
+      }
+      expect(parsed.axis).toEqual(expected)
+    })
+    it('add tick config if stepSize is specified in style', () => {
+      const chartElement = '#foo'
+      const settings: ChartSettings = {
+        data: [{ name: 'Foo', values: [1], axis: 'y2' }],
+        style: {
+          axis: {
+            y2: {
+              show: false,
+              stepSize: 5
+            },
+          },
+        },
+      }
+      const parsed = createOptions({ settings, chartElement })
+      const expected: Axis = {
+        y2: {
+          show: false,
+          tick: {
+            count: null,
+            stepSize: 5
+          }
+        },
+      }
+      expect(parsed.axis).toEqual(expected)
+    })
   })
   describe('createInfo', () => {
     let settings: ChartSettings