+++ libs/react/src/lib/dropdown/hooks.ts
@@ -45,6 +45,7 @@ interface HookResult {
 
 export const useDropdown = ({
   id,
+  value,
   texts,
   options,
   loop,
@@ -92,7 +93,7 @@ export const useDropdown = ({
       const newListItems: Props[] = dropdown.options.map((o) => ({
         ...(o.attributes as unknown as Props),
         className: o.classes?.join(' '),
-        children: o[dropdown.display],
+        children: o[dropdown.display] as string,
         selected: o.selected,
         onClick: () => handler?.select(o),
       }))
@@ -109,7 +110,7 @@ export const useDropdown = ({
           onClick: () => handler?.select(o, false),
         },
         spanProps: {
-          children: o[dropdown.display],
+          children: o[dropdown.display] as string,
         },
       }))
       const newMultiselect: MultiSelectProps = {
@@ -135,8 +136,20 @@ export const useDropdown = ({
   // When dropdown properties change
   useEffect(() => {
     if (!dropdown) return
-    handler?.update({ id, texts, options, loop, multiSelect })
-
+    handler?.update({
+      id,
+      value,
+      texts,
+      options,
+      loop,
+      multiSelect,
+      searchable,
+      searchFilter,
+      compareWith,
+      useValue,
+      display,
+      validator,
+    })
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [
     id,
@@ -165,6 +178,7 @@ export const useDropdown = ({
         createDropdown(
           {
             id,
+            value,
             texts,
             options,
             loop,